services/ingestors/trending_ingestor/requirements.txt
beautifulsoup4==4.12.2
requests==2.32.5
kafka-python==2.2.0
lxml==4.9.3
snscrape==0.7.0.20230622
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.2
feedparser==6.0.10